% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathmodelfit-package.R
\docType{package}
\name{pathmodelfit-package}
\alias{pathmodelfit-package}
\alias{_PACKAGE}
\title{pathmodelfit: Fit Indices for the Path Component of Latent Variable Structural Equation Models}
\description{
Compute fit indices for evaluating the path component of latent
    variable structural equation models. Available fit indices include 'RMSEA-P'
    and 'NSCI-P' originally presented and evaluated by Williams and O'Boyle (2011) 
    and demonstrated by O'Boyle and Williams (2011) and Williams, O'Boyle, & Yu (2020). 
    Also included are fit indices described by Hancock and Mueller (2011).
}
\examples{
# Load example data
data(mediationVC, package = "pathmodelfit")

# Define path model for lavaan
model4 <- "
Ldrrew =~ LdrrewI1 + LdrrewI2 + LdrrewI3
Jobcom =~ JobcomI1 + JobcomI2 + JobcomI3
Jobsat =~ JobsatI1 + JobsatI2 + JobsatI3
Orgcom =~ OrgcomI1 + OrgcomI2 + OrgcomI3
Jobsat ~ Ldrrew + Jobcom
Orgcom ~ Jobsat"

# Fit the pathmodel with lavaan's sem() function
fit <- lavaan::sem(model4, sample.cov = mediationVC, sample.nobs = 232)

# Compute the fit indices with the pathmodelfit() function
pathmodelfit_info = pathmodelfit::pathmodelfit(fit)

# View results
pathmodelfit_info
}
\references{
Hancock, G. R., & Mueller, R. O. (2011). The reliability paradox in assessing structural relations within covariance structure models. Educational and Psychological Measurement, 71(2), 306-324.

McNeish, D., & Hancock, G. R. (2018). The effect of measurement quality on targeted structural model fit indices: A comment on Lance, Beck, Fan, and Carter (2016). Psychological Methods, 23(1), 184–190. https://doi.org/10.1037/met0000157

O'Boyle, E. H., Jr., & Williams, L. J. (2011). Decomposing model fit: Measurement vs. theory in organizational research using latent variables. Journal of Applied Psychology, 96(1), 1–12. https://doi.org/10.1037/a0020539

Williams, L. J., & O’Boyle, E. H. (2011). The myth of global fit indices and alternatives for assessing latent variable relations. Organizational Research Methods, 14, 350-369.

Williams, L. J., O’Boyle, E. H., & Yu, J. (2020). Condition 9 and 10 tests of model confirmation: A review of James, Mulaik, and Brett (1982) and contemporary alternatives. Organizational Research Methods, 23, 1, 6-29.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/pathmodelfit}
  \item Report bugs at \url{https://github.com/tmsalab/pathmodelfit/issues}
}

}
\author{
\strong{Maintainer}: Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID})

Authors:
\itemize{
  \item Larry Williams \email{larry.williams@ttu.edu}
}

}
\keyword{internal}
