% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathmodelfit.R
\name{print.pathmodelfit}
\alias{print.pathmodelfit}
\title{Print a \code{pathmodelfit} object}
\usage{
\method{print}{pathmodelfit}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{pathmodelfit} object.}

\item{digits}{the minimum number of significant digits to be used: see
    \code{\link[base]{print.default}}.}

\item{...}{Additional parameters passed onto \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
Display results and invisibly return the \code{pathmodelfit} object if used in
a pipe.
}
\description{
Display a \code{pathmodelfit} object with fixed formatting.
}
\seealso{
\code{\link[=pathmodelfit]{pathmodelfit()}}
}
